pyproject.toml
[tool.poetry]
name = "wsi-service"
version = "0.13.4"
description = "RationAI & EMPAIA WSI Service to stream whole slide images"
authors = ["RationAI <469130@mail.muni.cz>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pydantic = {extras = ["dotenv"], version = "^2.0.0"}
fastapi = "^0.109.0"
requests = "^2.27.1"
uvicorn = "^0.27.0"
numpy = "^1.22.3"
Pillow = "^10.0.0"
imagecodecs = "^2024.0.0"
tifffile = "^2024.0.0"
uvloop = "^0.19.0"
aiohttp = "^3.8.1"
httptools = "^0.6.0"
filelock = "^3.6.0"
idna = "^3.3"
zipfly = "^6.0.3"
aiofiles = "^23.0.0"
httpx = "^0.26.0"
pydantic-settings = "^2.0.2"
six = "^1.16.0"
PyJWT = "^2.1.0"
cryptography = "^42.0.0"
# This is because
botocore = "==1.29.123"

[tool.poetry.dev-dependencies]
aioresponses = "^0.7.3"
black = "^24.0.0"
debugpy = "^1.5.1"
isort = "^5.10.1"
psutil = "^5.9.0"
pycodestyle = "^2.8.0"
pylint = "^3.0.0"
pytest-aioresponses = "^0.2.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.0.0"
pytest = "^8.0.0"
requests-mock = "^1.9.3"
wsi-service-plugin-openslide = { path = "wsi_service_base_plugins/openslide", develop = true }
wsi-service-plugin-pil = { path = "wsi_service_base_plugins/pil", develop = true }
wsi-service-plugin-tifffile = { path = "wsi_service_base_plugins/tifffile", develop = true }
wsi-service-plugin-tiffslide = { path = "wsi_service_base_plugins/tiffslide", develop = true }
wsi-service-plugin-wsidicom = { path = "wsi_service_base_plugins/wsidicom", develop = true }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
exclude = "wsi_service/models"

[tool.isort]
line_length = 120
profile = "black"
multi_line_output = 3
include_trailing_comma = true

[tool.pylint."MASTER"]
ignored-modules = ["pydantic"]
ignore = ["wsi_service/models/tests"]

[tool.pylint."MESSAGES CONTROL"]
max-line-length = 120
disable = "R,C,W0613,E0213,W0603,W0231,E1136,W0707,W0622,W0201,E0401,W3101,W0212,W0621"

[tool.pytest.ini_options]
addopts = "--asyncio-mode=strict"
testpaths = [
    "wsi_service/tests",
    "wsi_service_base_plugins/openslide/wsi_service_plugin_openslide/tests",
    "wsi_service_base_plugins/pil/wsi_service_plugin_pil/tests",
    "wsi_service_base_plugins/tifffile/wsi_service_plugin_tifffile/tests",
    "wsi_service_base_plugins/tiffslide/wsi_service_plugin_tiffslide/tests",
    "wsi_service_base_plugins/wsidicom/wsi_service_plugin_wsidicom/tests"
]
